smb_user_check_permission/requirements.txt
pysmb>=1.2.0
impacket>=0.9.27
smbprotocol>=1.5.0
cryptography>=3.4.8